requirements.txt
fastapi
uvicorn
langchain
langgraph
spacy
pandas
python-dotenv
httpx
groq
Pillow
aiofiles
uuid
pydantic